+++ packages/types/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update this changelog
 
+### Added
+- Added `isBigInt` and `ensureBigInt` functions
+
 ## [1.0.0] - 2024-07-07
 ### Added
 - Added `isDefined` and `ensureDefined` functions